fix handling of pressing /generate multiple times after paying

+++ main.py
@@ -54,21 +54,24 @@ async def echo(update: Update, context: ContextTypes.DEFAULT_TYPE):
 
 
 async def paid(update: Update, context: ContextTypes.DEFAULT_TYPE):
-    if payment.checkinvoice(user_state[update.effective_chat.id][1]['payment_hash']):
-        await context.bot.send_message(chat_id=update.effective_chat.id,
-                                       text="Generating pictures, this will take some time...")
-        file_paths = dalle.generate_and_download(user_state[update.effective_chat.id][0])
-        if isinstance(file_paths, list):
-            for n in file_paths:
-                await context.bot.send_photo(chat_id=update.effective_chat.id, photo=open(n, 'rb'))
-                os.remove(n)
+    if update.effective_chat.id in user_state:
+        if payment.checkinvoice(user_state[update.effective_chat.id][1]['payment_hash']):
+            await context.bot.send_message(chat_id=update.effective_chat.id,
+                                           text="Generating pictures, this will take some time...")
+            file_paths = dalle.generate_and_download(user_state[update.effective_chat.id][0])
+            if isinstance(file_paths, list):
+                for n in file_paths:
+                    await context.bot.send_photo(chat_id=update.effective_chat.id, photo=open(n, 'rb'))
+                    os.remove(n)
+            else:
+                await context.bot.send_message(chat_id=update.effective_chat.id,
+                                               text=messages.violation)
+            user_state.pop(update.effective_chat.id)
         else:
             await context.bot.send_message(chat_id=update.effective_chat.id,
-                                           text=messages.violation)
-        user_state.pop(update.effective_chat.id)
+                                           text="You haven't paid, press /generate again once you paid the invoice")
     else:
-        await context.bot.send_message(chat_id=update.effective_chat.id,
-                                       text="You haven't paid, press /generate again once you paid the invoice")
+        pass
 
 
 async def terms(update: Update, context: ContextTypes.DEFAULT_TYPE):
